refactor(sync): split sync pagination service into modular components
The SyncPaginationService has been refactored into multiple focused
helper classes to improve maintainability and separation of concerns.
The main service now delegates responsibilities to specialized helpers
for server pagination handling, DTO building, and progress tracking.
This change improves code organization and makes the service easier
to test and maintain.
