+++ src/lib/core/application/features/sync/sync_registration.dart
@@ -23,7 +23,7 @@ import 'package:whph/core/application/features/sync/services/sync_configuration_
 import 'package:whph/core/application/features/sync/services/sync_validation_service.dart';
 import 'package:whph/core/application/features/sync/services/sync_communication_service/sync_communication_service.dart';
 import 'package:whph/core/application/features/sync/services/sync_data_processing_service.dart';
-import 'package:whph/core/application/features/sync/services/sync_pagination_service.dart';
+import 'package:whph/core/application/features/sync/services/sync_pagination_service/sync_pagination_service.dart';
 import 'package:whph/core/application/features/sync/services/database_integrity_service.dart';
 import 'package:whph/infrastructure/persistence/shared/contexts/drift/drift_app_context.dart';
 import 'package:acore/acore.dart';
