+++ src/test/core/application/features/sync/services/sync_pagination_service_test.dart
@@ -6,7 +6,7 @@ import 'package:whph/core/application/features/sync/models/paginated_sync_data_d
 import 'package:whph/core/application/features/sync/models/sync_data.dart';
 import 'package:whph/core/application/features/sync/services/abstraction/i_sync_communication_service.dart';
 import 'package:whph/core/application/features/sync/services/abstraction/i_sync_configuration_service.dart';
-import 'package:whph/core/application/features/sync/services/sync_pagination_service.dart';
+import 'package:whph/core/application/features/sync/services/sync_pagination_service/sync_pagination_service.dart';
 import 'package:whph/core/domain/features/sync/sync_device.dart';
 import 'package:whph/core/domain/features/habits/habit_record.dart';
 import 'package:whph/core/application/shared/services/abstraction/i_repository.dart' as whph_repo;
